+++ test/unit-tests/config.test.ts
@@ -13,6 +13,7 @@ function createConfig(conf: Partial<ExtensionConfigurationSettings>): Configurat
         saveBeforeBuild: true,
         buildBeforeRun: true,
         clearOutputBeforeBuild: true,
+        colorizedBuildOutput: "off",
         configureSettings: {},
         cacheInit: null,
         preferredGenerators: [],